Separate minmax tree algorithm into its own library file.
Test it by comparing with base algorithm.
